requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-crispy-forms==1.4.0
django-profiles==0.2
django-registration==1.0
django-toolbelt==0.0.1
gdata==2.0.18
gunicorn==19.3.0
lxml==3.4.4
polib==1.0.7
psycopg2==2.6.1
python-docx==0.8.5
requests==2.7.0
South==1.0.2
static3==0.6.1
stripe==1.23.0
